refactor: complete rewrite (ES2015, currying, ...)
Rebuild from the ground up - modernize into ES2015 syntax. This means an environment supporting ES2015 features like `let`, `const`, arrow functions, etc. is necessary.

The format function is also now curried, meaning if you only provide the template string it will return a function that you can then call separately with your replacement parameters.

BREAKING CHANGE: Support for older runtimes has been dropped. An ES2015 environment is required.

BREAKING CHANGE: Replacements are no longer passed as an argument list to the format function. To allow for currying, replacement parameters should be passed as an Array if there is more than one. A single parameter does not need to be wrapped in an Array - it will be cast to one internally.