feat(methods): improve method invocation (#3)
* feat(methods): improve method invocation

Allow arguments with spaces and add the ability to nullify arguments.

BREAKING CHANGE: Invoking methods with arguments no longer space delimits arguments. Instead they are delimited by commas. Additionally, the `_` character can be used as a placeholder which is replaced by `null` when passed to the method being called.

* test(methods): test new argument parsing

* fix(syntax): missing comma in test file

* fix(tests): method invocation `_` escaping

Use a different function to test this that just returns the value passed to it.

* fix: remove duplicate tests

These were moved into the `method-invocation` tests.

* style: fix tabs -> spaces

Thanks a lot, GitHub editor.